+++ anomalib/data/inference.py
@@ -9,8 +9,7 @@ from typing import Any, Optional, Tuple, Union
 import albumentations as A
 from torch.utils.data.dataset import Dataset
 
-from anomalib.data.utils import get_image_filenames, read_image
-from anomalib.pre_processing import get_transforms
+from anomalib.data.utils import get_image_filenames, get_transforms, read_image
 
 
 class InferenceDataset(Dataset):
