+++ anomalib/data/utils/__init__.py
@@ -21,6 +21,7 @@ from .split import (
     random_split,
     split_by_label,
 )
+from .transform import get_transforms
 
 __all__ = [
     "generate_output_image_filename",
@@ -40,4 +41,5 @@ __all__ = [
     "masks_to_boxes",
     "boxes_to_masks",
     "boxes_to_anomaly_maps",
+    "get_transforms",
 ]
