+++ anomalib/pre_processing/__init__.py
@@ -5,6 +5,5 @@
 
 from .pre_process import PreProcessor
 from .tiler import Tiler
-from .transform import get_transforms
 
-__all__ = ["PreProcessor", "Tiler", "get_transforms"]
+__all__ = ["PreProcessor", "Tiler"]
